// src/types/betslip-selection.ts
interface BetslipSelection {
    eventId: number;
    outcomeId: number;
    eventGameId: number;
    outcomeOdds: number;
    outcomeName: string;
    eventName: string;
    eventGameName: string;
    sport: string;
    country: string;
    tournament: string;
    eventStart: string;
}

export { type BetslipSelection };